+++ src/core/Application.cpp
@@ -1,6 +1,9 @@
 #include "GDK/Application.hpp"
 
 #include <cstdlib>
+#include <stdexcept>
+
+#include <GL/glew.h>
 
 #include <glfwcxx/Core.hpp>
 
@@ -11,6 +14,7 @@ namespace gamedevkit {
 
 Application::Application()
     : details_{std::make_unique<Details>()}
+    , window_{nullptr}
 {
 }
 
@@ -21,15 +25,31 @@ Application::Application(int /*argc*/, char** /*argv[]*/)
 
 Application::~Application() = default;
 
-auto Application::window(std::unique_ptr<Window>&& /*window*/) -> Application&
+auto Application::window(std::unique_ptr<Window> window) -> Application&
 {
+    window_ = std::move(window);
     return *this;
 }
 
-auto Application::setup() -> void {}
+auto Application::setup() -> void
+{
+    if (nullptr == window_)
+        throw std::runtime_error("Cannot setup Application without Window being set");
+
+    window_->activate();
+
+    if (GLEW_OK != glewInit())
+        throw std::runtime_error("Failed to initialize GLEW");
+}
 
 auto Application::run() -> int
 {
+    while (false == window_->should_close()) {
+        window_->poll_events();
+
+        window_->swap_buffers();
+    }
+
     return EXIT_SUCCESS;
 }