+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 - Setup development infrastructure [(#1)](https://github.com/EugeneKuznetsov/gamedevkit/issues/1)
 - Setup continuous integration workflow [(#2)](https://github.com/EugeneKuznetsov/gamedevkit/issues/2)
+- Application class for setting up game infrastructure [(#7)](https://github.com/EugeneKuznetsov/gamedevkit/issues/7)
 
 ## Release